(i) moved to slice 2 for Diding's Saleh Bay model. (ii) Tried my hand at generalizing (again!) markets, I suppose it is that time of the year. I think it is now clear and easier to change prices on the fly; still not a fan of only the bins being accounted for and not the subdivision and I don't like the marginal price as it is structured but this will have to do. (iii)Also created an override to the map system to force certain cells to be at a specific altitude which was necessary here because there is a tiny 400m wide channel people use that kept getting turned into land by the pixellization. (iv) Finally, combined a Max Hours Out and Resting Time departing strategies together. Surely the flexible solution would be to have a generic factory for multiple Departing Strategies but the infrastructure has been there for many years and this may be the first time I actually needed it, so probably not that important.

+++ src/main/java/uk/ac/ox/oxfish/fisher/strategies/fishing/MaximumDaysDecorator.java
@@ -36,15 +36,15 @@ public class MaximumDaysDecorator implements FishingStrategy
 
     private FishingStrategy delegate;
 
-    private final int daysBeforeGoingHome;
+    private final double daysBeforeGoingHome;
 
-    public MaximumDaysDecorator(int daysBeforeGoingHome) {
+    public MaximumDaysDecorator(double daysBeforeGoingHome) {
         this(new FishUntilFullStrategy(1d),daysBeforeGoingHome);
 
 
     }
 
-    public MaximumDaysDecorator(FishingStrategy delegate, int daysBeforeGoingHome) {
+    public MaximumDaysDecorator(FishingStrategy delegate, double daysBeforeGoingHome) {
         this.delegate = delegate;
         this.daysBeforeGoingHome = daysBeforeGoingHome;
     }
@@ -93,7 +93,7 @@ public class MaximumDaysDecorator implements FishingStrategy
         return delegate;
     }
 
-    public int getDaysBeforeGoingHome() {
+    public double getDaysBeforeGoingHome() {
         return daysBeforeGoingHome;
     }
